refactor(p2): extract MessagePartsBlock into sibling page (stage 4 of 4)
Final cut. Moves the @OptIn(FlowPreview::class) MessagePartsBlock
(L320-L729 in trimmed Stage 3 coord, 410 lines) + its only-caller
helper streamingContentSize (L297-L298, with @Suppress) from
ChatMessage.kt into new sibling ChatMessageMessagePartsBlock.kt
(same package).

Byte-equal: bodies and signatures unchanged. @OptIn(FlowPreview::class)
preserved verbatim above @Composable. @Suppress("UNUSED_PARAMETER")
preserved on streamingContentSize.

Visibility:
- MessagePartsBlock: already widened to internal at Stage 3 (no change)
- streamingContentSize: unchanged (private, only same-sibling caller)

Coordinator strip: 729 -> 267 (-462 with import cleanup).
Imports trimmed: 46 orphans (Intent, BorderStroke, background, Box, Row,
padding, size, heightIn, widthIn, RoundedCornerShape, Icon, LaunchedEffect,
key, rememberUpdatedState, snapshotFlow, clip, drawWithContent, Offset,
Size, Color, HapticFeedbackType, LocalHapticFeedback, painterResource,
TextOverflow, fastAll, fastForEachIndexed, FileProvider, toFile, toUri,
FlowPreview, debounce, UIMessageAnnotation, HugeIcons, File02, MusicNote01,
Video01, R, AssistantAffectScope, replaceRegexes, GenerativeUiPlanner,
GenerativeWidgetParser, MarkdownBlock, ZoomableAsyncImage, workspaceColors,
shimmer, milliseconds).

Pre-existing dead preserved: getValue, setValue (used implicitly by `by`
delegation in coord at L137/L138 + L267/L268 of trimmed coord),
fastForEach (already 0 body refs in baseline), height (Modifier-extension
lesson #4: grep-pseudo-alive via Size constructor named arg + DrawScope.size.height
but 0 real Modifier.height(...) usage in baseline 3809e414).

Final coord size: 267 lines (from 1634, -83.7%).
- 44 imports
- hasRenderableChatMessageContent (internal)
- fun ChatMessage (public)
- animateContentSizeIf (internal)
- rememberChatMessageTextStyle (internal)

Sibling totals: ChatMessageCommon 156 + ChatMessageRenderers 316
+ ChatMessageVirtualItems 530 + ChatMessageMessagePartsBlock 480
= 1482 lines moved across 4 siblings.

Baseline 3/467 unit tests preserved.
